+++ buildTools/darwin-x64/gbdk/include/sms/sms.h
@@ -568,6 +568,18 @@ void set_tile_map_compat(uint8_t x, uint8_t y, uint8_t w, uint8_t h, const uint8
 #define set_bkg_tiles set_tile_map_compat
 #define set_win_tiles set_tile_map_compat
 
+extern uint8_t _map_tile_offset;
+inline void set_bkg_based_tiles(uint8_t x, uint8_t y, uint8_t w, uint8_t h, const uint8_t *tiles, uint8_t base_tile) {
+    _map_tile_offset = base_tile;
+    set_tile_map_compat(x, y, w, h, tiles);
+    _map_tile_offset = 0;
+}
+inline void set_win_based_tiles(uint8_t x, uint8_t y, uint8_t w, uint8_t h, const uint8_t *tiles, uint8_t base_tile) {
+    _map_tile_offset = base_tile;
+    set_tile_map_compat(x, y, w, h, tiles);
+    _map_tile_offset = 0;
+}
+
 void set_tile_submap(uint8_t x, uint8_t y, uint8_t w, uint8_t h, uint8_t map_w, const uint8_t *map) Z88DK_CALLEE PRESERVES_REGS(iyh, iyl);
 void set_tile_submap_compat(uint8_t x, uint8_t y, uint8_t w, uint8_t h, uint8_t map_w, const uint8_t *map) Z88DK_CALLEE PRESERVES_REGS(iyh, iyl);
 inline void set_bkg_submap(uint8_t x, uint8_t y, uint8_t w, uint8_t h, const uint8_t *map, uint8_t map_w) {
@@ -575,6 +587,18 @@ inline void set_bkg_submap(uint8_t x, uint8_t y, uint8_t w, uint8_t h, const uin
 }
 inline void set_win_submap(uint8_t x, uint8_t y, uint8_t w, uint8_t h, const uint8_t *map, uint8_t map_w) {
     set_tile_submap_compat(x, y, w, h, map_w, map);
+}
+
+extern uint8_t _submap_tile_offset;
+inline void set_bkg_based_submap(uint8_t x, uint8_t y, uint8_t w, uint8_t h, const uint8_t *map, uint8_t map_w, uint8_t base_tile) {
+    _submap_tile_offset = base_tile;
+    set_tile_submap_compat(x, y, w, h, map_w, map);
+    _submap_tile_offset = 0;
+}
+inline void set_win_based_submap(uint8_t x, uint8_t y, uint8_t w, uint8_t h, const uint8_t *map, uint8_t map_w, uint8_t base_tile) {
+    _submap_tile_offset = base_tile;
+    set_tile_submap_compat(x, y, w, h, map_w, map);
+    _submap_tile_offset = 0;
 }
 
 void fill_rect(uint8_t x, uint8_t y, uint8_t w, uint8_t h, const uint16_t tile) Z88DK_CALLEE PRESERVES_REGS(iyh, iyl);