+++ buildTools/linux-x64/gbdk/include/asm/types.h
@@ -4,14 +4,12 @@
 #ifndef ASM_TYPES_INCLUDE
 #define ASM_TYPES_INCLUDE
 
-#ifdef __PORT_gbz80
-  #include <asm/gbz80/types.h>
+#if defined(__PORT_gbz80)
+#include <asm/gbz80/types.h>
+#elif defined(__PORT_z80)
+#include <asm/z80/types.h>
 #else
-  #ifdef __PORT_z80
-    #include <asm/z80/types.h>
-  #else
-    #error Unrecognised port
-  #endif
+#error Unrecognised port
 #endif
 
 #ifndef OLDCALL
@@ -20,6 +18,18 @@
 #else
 #define OLDCALL
 #endif
+#endif
+
+#ifdef __SDCC
+#define PRESERVES_REGS(...) __preserves_regs(__VA_ARGS__)
+#define NAKED   __naked
+#define SFR     __sfr
+#define AT(A)   __at(A)
+#else
+#define PRESERVES_REGS(...)
+#define NAKED
+#define SFR
+#define AT(A)
 #endif
 
 #ifndef NONBANKED
@@ -38,34 +48,26 @@
 /** TRUE or FALSE.
     @anchor file_asm_types_h
  */
-typedef INT8		BOOLEAN;
-
-#if BYTE_IS_UNSIGNED
-
-typedef UINT8		BYTE;
-typedef UINT16		WORD;
-typedef UINT32		DWORD;
-
-#else
+typedef INT8    BOOLEAN;
 
 /** Signed 8 bit.
  */
-typedef INT8         	BYTE;
+typedef INT8    BYTE;
 /** Unsigned 8 bit.
  */
-typedef UINT8        	UBYTE;
+typedef UINT8   UBYTE;
 /** Signed 16 bit */
-typedef INT16      	WORD;
+typedef INT16   WORD;
 /** Unsigned 16 bit */
-typedef UINT16       	UWORD;
+typedef UINT16  UWORD;
 /** Signed 32 bit */
-typedef INT32       	LWORD;
+typedef INT32   LWORD;
 /** Unsigned 32 bit */
-typedef UINT32      	ULWORD;
+typedef UINT32  ULWORD;
 /** Signed 32 bit */
-typedef INT32	   	DWORD;
+typedef INT32	  DWORD;
 /** Unsigned 32 bit */
-typedef UINT32	   	UDWORD;
+typedef UINT32	UDWORD;
 
 /** Useful definition for working with 8 bit + 8 bit fixed point values
 
@@ -74,17 +76,15 @@ typedef UINT32	   	UDWORD;
     Use `.b.h` and `.b.l` (or just `.h` and `.l`) to directly access it's high and low unsigned 8 bit values.
  */
 typedef union _fixed {
-  struct {
-    UBYTE l;
-    UBYTE h;
-  };
-  struct {
-    UBYTE l;
-    UBYTE h;
-  } b;
-  UWORD w;
+    struct {
+        UBYTE l;
+        UBYTE h;
+    };
+    struct {
+        UBYTE l;
+        UBYTE h;
+    } b;
+    UWORD w;
 } fixed;
-
-#endif
 
 #endif