+++ src/lib/compiler/compileData.js
@@ -61,6 +61,7 @@ import {
   compileSceneProjectiles,
   compileSceneProjectilesHeader,
   spriteSheetSymbol,
+  compileSaveSignature,
 } from "./compileData2";
 import compileSGBImage from "./sgb";
 import { readFileToTilesData } from "../tiles/tileData";
@@ -70,6 +71,7 @@ import { compileMusicTracks, compileMusicHeader } from "./compileMusic";
 import { chunk } from "../helpers/array2";
 import { toProjectileHash } from "./scriptBuilder";
 import {
+  calculateAutoFadeEventIdDenormalised,
   walkDenormalizedSceneEvents,
   walkDenormalizedScenesEvents,
 } from "lib/helpers/eventHelpers";
@@ -132,8 +134,11 @@ export const precompileStrings = (scenes, customEventsLookup) => {
   walkDenormalizedScenesEvents(
     scenes,
     {
-      customEventsLookup,
-      maxDepth: MAX_NESTED_SCRIPT_DEPTH,
+      customEvents: {
+        lookup: customEventsLookup,
+        maxDepth: MAX_NESTED_SCRIPT_DEPTH,
+        args: {},
+      },
     },
     (cmd) => {
       if (
@@ -178,8 +183,11 @@ export const precompileBackgrounds = async (
   walkDenormalizedScenesEvents(
     scenes,
     {
-      customEventsLookup,
-      maxDepth: MAX_NESTED_SCRIPT_DEPTH,
+      customEvents: {
+        lookup: customEventsLookup,
+        maxDepth: MAX_NESTED_SCRIPT_DEPTH,
+        args: {},
+      },
     },
     (cmd) => {
       if (eventHasArg(cmd, "backgroundId")) {
@@ -476,8 +484,11 @@ export const precompileSprites = async (
   walkDenormalizedScenesEvents(
     scenes,
     {
-      customEventsLookup,
-      maxDepth: MAX_NESTED_SCRIPT_DEPTH,
+      customEvents: {
+        lookup: customEventsLookup,
+        maxDepth: MAX_NESTED_SCRIPT_DEPTH,
+        args: {},
+      },
     },
     (event) => {
       if (event.args) {
@@ -555,8 +566,11 @@ export const precompileAvatars = async (
   walkDenormalizedScenesEvents(
     scenes,
     {
-      customEventsLookup,
-      maxDepth: MAX_NESTED_SCRIPT_DEPTH,
+      customEvents: {
+        lookup: customEventsLookup,
+        maxDepth: MAX_NESTED_SCRIPT_DEPTH,
+        args: {},
+      },
     },
     (event) => {
       if (event.args) {
@@ -597,8 +611,11 @@ export const precompileEmotes = async (
   walkDenormalizedScenesEvents(
     scenes,
     {
-      customEventsLookup,
-      maxDepth: MAX_NESTED_SCRIPT_DEPTH,
+      customEvents: {
+        lookup: customEventsLookup,
+        maxDepth: MAX_NESTED_SCRIPT_DEPTH,
+        args: {},
+      },
     },
     (event) => {
       if (event.args) {
@@ -640,8 +657,11 @@ export const precompileMusic = (
   walkDenormalizedScenesEvents(
     scenes,
     {
-      customEventsLookup,
-      maxDepth: MAX_NESTED_SCRIPT_DEPTH,
+      customEvents: {
+        lookup: customEventsLookup,
+        maxDepth: MAX_NESTED_SCRIPT_DEPTH,
+        args: {},
+      },
     },
     (cmd) => {
       if (
@@ -723,8 +743,11 @@ export const precompileFonts = async (
   walkDenormalizedScenesEvents(
     scenes,
     {
-      customEventsLookup,
-      maxDepth: MAX_NESTED_SCRIPT_DEPTH,
+      customEvents: {
+        lookup: customEventsLookup,
+        maxDepth: MAX_NESTED_SCRIPT_DEPTH,
+        args: {},
+      },
     },
     (cmd) => {
       if (cmd.args && cmd.args.fontId !== undefined) {
@@ -842,8 +865,11 @@ export const precompileScenes = (
     walkDenormalizedSceneEvents(
       scene,
       {
-        customEventsLookup,
-        maxDepth: MAX_NESTED_SCRIPT_DEPTH,
+        customEvents: {
+          lookup: customEventsLookup,
+          maxDepth: MAX_NESTED_SCRIPT_DEPTH,
+          args: {},
+        },
       },
       (event, _scene, actor, _trigger) => {
         if (
@@ -1114,6 +1140,7 @@ const compile = async (
   const isColor = customColorsEnabled || isSGB;
 
   const precompiledEngineFields = precompileEngineFields(engineFields);
+  const customEventsLookup = keyBy(projectData.customEvents, "id");
 
   // Add UI data
   output["frame_image.c"] = compileFrameImage(precompiled.frameTiles);
@@ -1200,7 +1227,11 @@ const compile = async (
 
       const scriptName = `script_s${sceneIndex}${entityCode}_${scriptTypeCode}`;
 
-      if (script.length === 0) {
+      if (
+        script.length === 0 &&
+        // Generate scene init for empty script if autoFade is not disabled
+        (scriptTypeCode !== "init" || scene.autoFadeSpeed === null)
+      ) {
         return null;
       }
 
@@ -1232,7 +1263,6 @@ const compile = async (
         warnings,
         loop,
         lock,
-        init: scriptTypeCode === "init",
         engineFields: precompiledEngineFields,
         output: [],
         additionalScripts,
@@ -1290,6 +1320,28 @@ VM_ACTOR_SET_SPRITESHEET_BY_REF .ARG2, .ARG1`,
           [scene.script] || []
         )
         .flat();
+
+      // Inject automatic Scene Fade In if required
+      if (scene.autoFadeSpeed !== null) {
+        const autoFadeId = calculateAutoFadeEventIdDenormalised(
+          initScript,
+          customEventsLookup
+        );
+        const autoFadeIndex = initScript.findIndex(
+          (item) => item.id === autoFadeId
+        );
+        const fadeEvent = {
+          command: "EVENT_FADE_IN",
+          args: {
+            speed: scene.autoFadeSpeed,
+          },
+        };
+        if (autoFadeIndex > -1) {
+          initScript.splice(autoFadeIndex, 0, fadeEvent);
+        } else if (autoFadeId !== "MANUAL") {
+          initScript.push(fadeEvent);
+        }
+      }
 
       // Compile scene start script
       return compileScript(
@@ -1680,6 +1732,10 @@ VM_ACTOR_SET_SPRITESHEET_BY_REF .ARG2, .ARG1`,
       })
       .join("") +
     `#endif\n`;
+
+  output[`game_signature.c`] = compileSaveSignature(
+    JSON.stringify(projectData)
+  );
 
   return {
     files: output,