+++ src/lib/compiler/compileEntityEvents.js
@@ -1,4 +1,3 @@
-import { EVENT_FADE_IN } from "./eventTypes";
 import ScriptBuilder from "./scriptBuilder";
 
 const STRING_NOT_FOUND = "STRING_NOT_FOUND";
@@ -16,7 +15,6 @@ const compileEntityEvents = (scriptName, input = [], options = {}) => {
     warnings,
     loop,
     lock,
-    init,
     isFunction,
   } = options;
 
@@ -35,15 +33,13 @@ const compileEntityEvents = (scriptName, input = [], options = {}) => {
     }),
   };
 
-  let globalHasInit = false;
+  let hasInit = false;
 
   const compileEventsWithScriptBuilder = (
     scriptBuilder,
     subInput = [],
     isBranch = false
   ) => {
-    let hasInit = false;
-
     // eslint-disable-next-line global-require
     // eslint-disable-next-line @typescript-eslint/no-var-requires
     const events = require("../events").default;
@@ -55,24 +51,6 @@ const compileEntityEvents = (scriptName, input = [], options = {}) => {
         continue;
       }
       if (events[command]) {
-        if (
-          init &&
-          !hasInit &&
-          !globalHasInit &&
-          events[command].waitUntilAfterInitFade
-        ) {
-          // Found an event that cannot happen before init fade in
-          if (i > 0 || isBranch) {
-            // Force await the next frame if not the first command
-            // at top level so that actors update before fade in occurs
-            scriptBuilder.nextFrameAwait();
-          }
-          if (command !== EVENT_FADE_IN) {
-            scriptBuilder.fadeIn();
-          }
-          hasInit = true;
-          globalHasInit = true;
-        }
         try {
           events[command].compile(
             { ...subInput[i].args, ...subInput[i].children },
@@ -82,13 +60,6 @@ const compileEntityEvents = (scriptName, input = [], options = {}) => {
               event: subInput[i],
             }
           );
-          if (!isBranch && !hasInit && globalHasInit) {
-            // A branch caused a fade in, but hasn't faded in globally yet
-            // Need to force fade in case one path of branch didn't already fade
-            scriptBuilder.nextFrameAwait();
-            scriptBuilder.fadeIn();
-            hasInit = true;
-          }
         } catch (e) {
           console.error(e);
           throw new Error(
@@ -142,15 +113,6 @@ const compileEntityEvents = (scriptName, input = [], options = {}) => {
         scriptBuilder.nextFrameAwait();
         scriptBuilder.labelGoto(loopId);
       }
-
-      if (init && !globalHasInit) {
-        // No part of script caused a fade in so do this
-        // before ending the script
-        scriptBuilder.nextFrameAwait();
-        scriptBuilder.fadeIn();
-        globalHasInit = true;
-      }
-
       if (isFunction) {
         if (scriptBuilder.includeActor) {
           scriptBuilder.stackPtr += 4;