+++ src/lib/events/eventInputAwait.js
@@ -29,4 +29,5 @@ module.exports = {
   groups,
   fields,
   compile,
+  waitUntilAfterInitFade: true,
 };