+++ src/lib/events/eventInputScriptSet.js
@@ -53,4 +53,5 @@ module.exports = {
   groups,
   fields,
   compile,
+  allowChildrenBeforeInitFade: true,
 };