+++ src/lib/helpers/8bit.ts
@@ -1,5 +1,7 @@
 export const wrap8Bit = (val: number) => (256 + (val % 256)) % 256;
 export const wrap16Bit = (val: number) => (65536 + (val % 65536)) % 65536;
+export const wrap32Bit = (val: number) =>
+  (4294967296 + (val % 4294967296)) % 4294967296;
 
 export const decBin = (dec: number) =>
   wrap8Bit(dec).toString(2).padStart(8, "0");
@@ -15,6 +17,9 @@ export const decHex16 = (dec: number) =>
 
 export const decHex16Val = (dec: number) =>
   wrap16Bit(dec).toString(16).padStart(4, "0").toUpperCase();
+
+export const decHex32Val = (dec: number) =>
+  wrap32Bit(dec).toString(16).padStart(8, "0").toUpperCase();
 
 export const decOct = (dec: number) =>
   wrap8Bit(dec).toString(8).padStart(3, "0");