+++ test/data/compiler/compileEntityEvents.test.js
@@ -21,6 +21,7 @@ test("should compile empty events", () => {
 
 ___bank_testname = 255
 .globl ___bank_testname
+.CURRENT_SCRIPT_BANK == ___bank_testname
 
 _testname::
         ; Stop Script
@@ -48,6 +49,7 @@ test("should collapse multiple end events", () => {
 
 ___bank_testname = 255
 .globl ___bank_testname
+.CURRENT_SCRIPT_BANK == ___bank_testname
 
 _testname::
         ; Stop Script
@@ -80,6 +82,7 @@ test("should output text command", async () => {
 
 ___bank_testname = 255
 .globl ___bank_testname
+.CURRENT_SCRIPT_BANK == ___bank_testname
 
 _testname::
         ; Text Dialogue
@@ -133,6 +136,7 @@ test("should output text with avatar command", async () => {
 
 ___bank_testname = 255
 .globl ___bank_testname
+.CURRENT_SCRIPT_BANK == ___bank_testname
 
 _testname::
         ; Text Dialogue
@@ -197,6 +201,7 @@ test("should allow conditional statements", async () => {
 
 ___bank_testname = 255
 .globl ___bank_testname
+.CURRENT_SCRIPT_BANK == ___bank_testname
 
 _testname::
         ; If Variable True
@@ -282,6 +287,7 @@ test("should allow commands after conditional", async () => {
 
 ___bank_testname = 255
 .globl ___bank_testname
+.CURRENT_SCRIPT_BANK == ___bank_testname
 
 _testname::
         ; If Variable True