+++ test/data/compiler/scriptBuilder2.test.ts
@@ -128,6 +128,7 @@ ACTOR = -4
 
 ___bank_MY_SCRIPT = 255
 .globl ___bank_MY_SCRIPT
+.CURRENT_SCRIPT_BANK == ___bank_MY_SCRIPT
 
 _MY_SCRIPT::
         ; Local Actor
@@ -176,6 +177,7 @@ test("Should be able to open dialogue boxes", async () => {
 
 ___bank_MY_SCRIPT = 255
 .globl ___bank_MY_SCRIPT
+.CURRENT_SCRIPT_BANK == ___bank_MY_SCRIPT
 
 _MY_SCRIPT::
         ; Text Dialogue
@@ -243,6 +245,7 @@ test("Should be able to conditionally execute if variable is true with event arr
 
 ___bank_MY_SCRIPT = 255
 .globl ___bank_MY_SCRIPT
+.CURRENT_SCRIPT_BANK == ___bank_MY_SCRIPT
 
 _MY_SCRIPT::
         ; If Variable True
@@ -307,6 +310,7 @@ test("Should be able to conditionally execute if variable is true with function 
 
 ___bank_MY_SCRIPT = 255
 .globl ___bank_MY_SCRIPT
+.CURRENT_SCRIPT_BANK == ___bank_MY_SCRIPT
 
 _MY_SCRIPT::
         ; If Variable True
@@ -399,6 +403,7 @@ test("Should be able to conditionally execute if variable is true with nested fu
 
 ___bank_MY_SCRIPT = 255
 .globl ___bank_MY_SCRIPT
+.CURRENT_SCRIPT_BANK == ___bank_MY_SCRIPT
 
 _MY_SCRIPT::
         ; If Variable True
@@ -490,6 +495,7 @@ test("Should be able to define labels and jump", () => {
 
 ___bank_MY_SCRIPT = 255
 .globl ___bank_MY_SCRIPT
+.CURRENT_SCRIPT_BANK == ___bank_MY_SCRIPT
 
 _MY_SCRIPT::
 1$: